Day 10: Ugly
Well this has a number of issues. It has an implementation of
find_fewest_presses that overflows the heap. Then it has one that
doesn't, but isn't as efficient as it could be. Finally, it contains a
bumbling attempt to use Z3 to solve part 2. This *works* but ooooooh boy
is it ugly: type assertions, ignoring errors, runtime messages from the
WASM build of Z3, oh and it's slow.